fix!: Remove double indirection in `MoveFieldLayout` (#21249) (#5740)
The `MoveFieldLayout` struct `fields` was wrapped in a box as part of https://github.com/MystenLabs/sui/pull/19310, which states in its reasoning:

> Boxes fields in the MoveTypeLayout to make the enum size smaller for type layouts.

However, this field being boxed does not affect the enums which this was referring to as it is also boxed in the variants. This means that the variant size of those enums is always the size of a pointer. Thus, this
additional box around the vec serves only to create an additional indirection.