hotfix(prettier-move): update lock files (#6071)
# Description of change

Please write a summary of your changes and why you made them.

## Links to any relevant issues

Fixes #6070 

## Type of change

- Enhancement

## How the change has been tested

Describe the tests that you ran to verify your changes.

Make sure to provide instructions for the maintainer as well as any
relevant configurations.

## Change checklist

Tick the boxes that are relevant to your changes, and delete any items
that are not.

- [x] I have followed the contribution guidelines for this project
- [x] I have performed a self-review of my own code
- [ ] I have commented my code, particularly in hard-to-understand areas
- [ ] I have made corresponding changes to the documentation
- [ ] I have added tests that prove my fix is effective or that my
feature works
- [ ] I have checked that new and existing unit tests pass locally with
my changes

Co-authored-by: Marc Espin <mespinsanz@gmail.com>
Co-authored-by: evavirseda <evirseda@boxfish.studio>