ci: do not run workflows in draft, except keyword is available in PR description (#6318)
# Description of change

This PR should stop running workflows if the PR is a draft. This should
reduce the load on the CI.
Sometimes the developer wants to run the CI anyway, but keep the PR as a
draft.
For this the magic keyword `[run-ci]` should be added in the PR
description.

## Type of change

- Enhancement (a non-breaking change which adds functionality)

## How the change has been tested

The changes seem to work fine.
First the keyword was not present in the PR description, and the tests
were skipped.
Then I changed the PR description to add the keyword and pushed another
commit (force-push). After that the tests were executed.