feat(explorer): add toogle to filter between system and user txs (#6317)
# Description of change

:warning:  Nested PR
Add a checkbox to filter between system and user tx in explorer

## Links to any relevant issues

fixes #4999 

## Type of change

Choose a type of change, and delete any options that are not relevant.

- Enhancement (a non-breaking change which adds functionality)

## How the change has been tested

Describe the tests that you ran to verify your changes.

Make sure to provide instructions for the maintainer as well as any
relevant configurations.

## Change checklist

Tick the boxes that are relevant to your changes, and delete any items
that are not.

- [ ] I have followed the contribution guidelines for this project
- [ ] I have performed a self-review of my own code
- [ ] I have commented my code, particularly in hard-to-understand areas
- [ ] I have made corresponding changes to the documentation
- [ ] I have added tests that prove my fix is effective or that my
feature works
- [ ] I have checked that new and existing unit tests pass locally with
my changes

---------

Co-authored-by: evavirseda <evirseda@boxfish.studio>
Co-authored-by: Marc Espin <mespinsanz@gmail.com>