feat(wallet-dashboard): fix toasts and dialogs overlap (#6347)
# Description of change
now:

![image](https://github.com/user-attachments/assets/15064f3d-5e7b-44c8-bf35-28f9bc084a27)
before:

![image](https://github.com/user-attachments/assets/66179977-687a-4e84-ac7f-f7489cf67f68)


## Links to any relevant issues

fixes https://github.com/iotaledger/iota/issues/6344

## Type of change

Choose a type of change, and delete any options that are not relevant.

- Bug fix (a non-breaking change which fixes an issue)
- Enhancement (a non-breaking change which adds functionality)
- Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- Documentation Fix

## How the change has been tested

Describe the tests that you ran to verify your changes.

Make sure to provide instructions for the maintainer as well as any
relevant configurations.

## Change checklist

Tick the boxes that are relevant to your changes, and delete any items
that are not.

- [ ] I have followed the contribution guidelines for this project
- [ ] I have performed a self-review of my own code
- [ ] I have commented my code, particularly in hard-to-understand areas
- [ ] I have made corresponding changes to the documentation
- [ ] I have added tests that prove my fix is effective or that my
feature works
- [ ] I have checked that new and existing unit tests pass locally with
my changes