ci: fix behavior where env vars are set but empty (#6485)
# Description of change

The checks in bash for unset env vars were not correct, because they
were assigned to local variables to be able to pass them to other
functions. This was now fixed with helper variables `xxx_given`.

With this fix, the CI shouldn't search for changed crates itself
anymore, if the given "changed crates" list was empty.

## Type of change

- Bug fix (a non-breaking change which fixes an issue)

## How the change has been tested

I tested it locally by setting and unsetting the variables and checking
the output. It seems to work as expected now.

### Release Notes

<!--
Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.
-->

- [ ] Protocol:
- [ ] Nodes (Validators and Full nodes):
- [ ] Indexer:
- [ ] JSON-RPC:
- [ ] GraphQL:
- [ ] CLI:
- [ ] Rust SDK:
- [ ] REST API: