fix(iota-framework-snapshot): update snapshot loading (#6478)
# Description of change

This is a pull of the following upstream change:

```
Since deployed protocol versions are not guaranteed to be sequential,
snapshots are not guaranteed to be sequential (e.g., 45 may exist, 46
may not, but 47 may).

Previously when trying to load a snapshot for a protocol version we
would only load a snapshot if there was one for that exact version,
otherwise we'd default to the newest version of the framework.

This updates snapshot loading so if a snapshot for a particular version
`v` is not found, the least version greater than the given version is
returned.
```

## Links to any relevant issues

fixes #6236