feat(cli): store framework versions in iota binary (#6438)
[run-ci]
# Description of change

Upstream pull, adds `framework_versions.rs` which allows getting
metadata from historical framework versions. This information is
gathered at build time from the manifest in the
`iota-framework-snapshot` crate. Additional data was added to that
manifest.

## Links to any relevant issues

fixes #6220 

## Type of change

Choose a type of change, and delete any options that are not relevant.

- Enhancement (a non-breaking change which adds functionality)

## How the change has been tested

Describe the tests that you ran to verify your changes.

Make sure to provide instructions for the maintainer as well as any
relevant configurations.

## Change checklist

Tick the boxes that are relevant to your changes, and delete any items
that are not.

- [ ] I have followed the contribution guidelines for this project
- [ ] I have performed a self-review of my own code
- [ ] I have commented my code, particularly in hard-to-understand areas
- [ ] I have made corresponding changes to the documentation
- [ ] I have added tests that prove my fix is effective or that my
feature works
- [ ] I have checked that new and existing unit tests pass locally with
my changes