scripts: fix cargo sort grouping only in deps sections (#6533)
# Description of change

the cargo sort script added internal and external groups also in
sections that contained entries that match the internal crate names.
this was the case in the "features" section and was now fixed.

## Type of change

- Bug fix (a non-breaking change which fixes an issue)

## How the change has been tested

I cherry picked the commit on top of the `iota-names` branch and
executing cargo sort yielded the correct result.

## Change checklist

- [x] I have followed the contribution guidelines for this project
- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas
- [ ] I have made corresponding changes to the documentation
- [ ] I have added tests that prove my fix is effective or that my
feature works
- [ ] I have checked that new and existing unit tests pass locally with
my changes