feat(iota): improve name `register` command (#6525)
# Description of change

In https://github.com/iotaledger/iota-private/issues/140 it was
suggested that we could allow users to immediately set the target
address and reverse lookup when registering, so this PR adds those flags
to the command.

## Type of change

Choose a type of change, and delete any options that are not relevant.

- Enhancement (a non-breaking change which adds functionality)

## How the change has been tested



## Change checklist

Tick the boxes that are relevant to your changes, and delete any items
that are not.

- [ ] I have followed the contribution guidelines for this project
- [ ] I have performed a self-review of my own code
- [ ] I have commented my code, particularly in hard-to-understand areas
- [ ] I have made corresponding changes to the documentation
- [ ] I have added tests that prove my fix is effective or that my
feature works
- [ ] I have checked that new and existing unit tests pass locally with
my changes

---------

Co-authored-by: Thibault Martinez <thibault@iota.org>