feat: Add generalized method for getting all pages (#6697)
# Description of change

When calling pages endpoints, it can be difficult to write the
boilerplate around managing the pages and cursor. This PR adds a
generalized method for getting all of the pages as a stream which can be
collected.

### Release Notes

<!--
Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.
-->

- [ ] Protocol:
- [ ] Nodes (Validators and Full nodes):
- [ ] Indexer:
- [ ] JSON-RPC:
- [ ] GraphQL:
- [ ] CLI:
- [x] Rust SDK: Add `PagedFn` trait which can be used to more easily
manage paged API calls.
- [ ] REST API:

---------

Co-authored-by: Thibault Martinez <thibault@iota.org>