feat(protocol-research): sequencer updates (#5763)
# Description of change

Upgrade the sequencer to use the "shelf-stacking" approach to make
better use of parallel execution during congestion.
The sequencing algorithm prior to this PR, roughly speaking, does the
following upon each consensus commit:

Sort transactions according to gas price.
Evaluate transactions one by one and decide whether they can be
scheduled for execution or should be deferred to the next consensus
commit. A transaction is deferred if it wants to write to a so-called
congested object.
This PR introduces a new decision logic for the second step. Deferments
are less likely, because even if the tx touches a congested object, it
is checked whether it could be executed in parallel with previously
scheduled transactions.

Consider the following example:

Tx 1 writes to objects A and B,
Tx 2 writes to objects B and C,
Tx 3 write to objects C and D

Assume congestion limit is set to 2. The current sequencing algorithm
will schedule Tx1 and Tx2 but defer Tx3 because objects B and C are
considered "congested" once Tx2 is scheduled, and Tx3 wants to write to
a congested object. The newly introduced algorithm will detect that Tx3
could be executed in parallel with Tx1, so Tx3 will be scheduled for
execution, with the condition that Tx3 must be executed before Tx2.

This PR introduces the concept of "slots" to implement this logic. A
slot represents an imaginary timeframe that indicates when an object was
allocated to a transaction.

The new algorithm schedules at least as many transactions as the
existing one but sacrifices strict gas price ordering. However, it still
guarantees that two identical transactions (i.e., those with the same
input objects) will be ordered by gas price.

This update will initially be enabled only in Devnet, and will be enabled in Testnet/Mainnet following stress testing in Devnet.

## Links to any relevant issues

closes #5744

## Type of change

- Enhancement (a non-breaking change which adds functionality)

## How the change has been tested

Existing tests pass

## Change checklist

Tick the boxes that are relevant to your changes, and delete any items
that are not.

- [x] I have followed the contribution guidelines for this project
- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas
- [ ] I have made corresponding changes to the documentation
- [x] I have added tests that prove my fix is effective or that my
feature works
- [x] I have checked that new and existing unit tests pass locally with
my changes

---------

Co-authored-by: cuileri <canumutileri@gmail.com>
Co-authored-by: Roman Overko <roman.overko@iota.org>