feat(iota-indexer): add `reliability.md` (#6694)
# Description of change

To enable better monitoring with Grafana for catching regressions and
performance issues, this PR expands on the reliability requirements for
the `iota-indexer`.

## Links to any relevant issues

fixes #6374 

## Type of change

- Enhancement (a non-breaking change which adds functionality)

## How the change has been tested

No code change was done so as test only basic lists & formatters were
used like `dprint fmt` when creating the file.

> [!NOTE]
> The MD file does not influence the actual Indexer performance, no
further tests were performed.

## Change checklist

Tick the boxes that are relevant to your changes, and delete any items
that are not.

- [x] I have followed the contribution guidelines for this project
- [x] I have performed a self-review of my own code

### Release Notes

<!--
Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.
-->

- [ ] Protocol:
- [ ] Nodes (Validators and Full nodes):
- [ ] Indexer:
- [ ] JSON-RPC:
- [ ] GraphQL:
- [ ] CLI:
- [ ] Rust SDK:
- [ ] REST API:

---------

Co-authored-by: Konstantinos Demartinos <konstantinos.demartinos@iota.org>