fix(iota-data-ingestion-core): make file updates atomic (#6918)
# Description of change

This PR refactors the `FileProgressStore` to ensure atomic updates when
storing watermark data.
It leverages a temporary file to write the new data, and then uses an
atomic rename operation to replace the original file with the updated
version.
  
This approach guarantees that the watermark file is never left in a
partially written or corrupted state, even in the event of a crash or
unexpected shutdown, ensuring data integrity and reliability.

## Links to any relevant issues

fixes #6889 

## Type of change


- Bug fix (a non-breaking change which fixes an issue)

## How the change has been tested

- Added 2 test cases to verify the results.
 ```shell
 cargo test -p iota-data-ingestion-core
```
Make sure to provide instructions for the maintainer as well as any relevant configurations.

- [x] Basic tests (linting, compilation, formatting, unit/integration tests)
    - Existing tests passed 
- [x] Patch-specific tests (correctness, functionality coverage)
    - Added 2 tests in the `iota-data-ingestion-core` crate that verifies the actual patch changes
    -  `file_progress_store_save_timeout_simulates_crash` & `file_progress_store`

### Infrastructure QA (only required for crates that are maintained by @iotaledger/infrastructure)

- [x] Verification of API backward compatibility.

> [!NOTE]
> The other checkboxes were removed since this patch does not directly affect the indexer performance, nevertheless running locally the ingestion framework for checkpoint processing risulted in the same performance as before, not visible regression were spotted.

## Change checklist

Tick the boxes that are relevant to your changes, and delete any items that are not.

- [x] I have followed the contribution guidelines for this project
- [x] I have performed a self-review of my own code
- [x] I have added tests that prove my fix is effective or that my feature works
- [x] I have checked that new and existing unit tests pass locally with my changes

### Release Notes

<!--
Check each box that your changes affect. If none of the boxes relate to your changes, release notes aren't required.

For each box you select, include information after the relevant heading that describes the impact of your changes that a user might notice and any actions they must take to implement updates.
-->

- [ ] Protocol:
- [ ] Nodes (Validators and Full nodes):
- [ ] Indexer:
- [ ] JSON-RPC:
- [ ] GraphQL:
- [ ] CLI:
- [ ] Rust SDK:
- [ ] REST API: